+++ crates/zizmor/tests/integration/audit/dangerous_triggers.rs
@@ -18,6 +18,75 @@ fn test_actions_labeler_exception() -> anyhow::Result<()> {
 }
 
 #[test]
+fn test_dangerous_trigger_bare_style() -> anyhow::Result<()> {
+    insta::assert_snapshot!(
+    zizmor()
+        .input(input_under_test(
+            "dangerous-triggers/dangerous-bare.yml"
+        ))
+        .run()?,
+    @"
+    error[dangerous-triggers]: use of fundamentally insecure workflow trigger
+     --> @@INPUT@@:1:1
+      |
+    1 | on: pull_request_target
+      | ^^^^^^^^^^^^^^^^^^^^^^^ pull_request_target is almost always used insecurely
+      |
+      = note: audit confidence → Medium
+
+    4 findings (3 suppressed): 0 informational, 0 low, 0 medium, 1 high
+    "
+    );
+
+    Ok(())
+}
+
+/// Ensures that we produce a reasonable diagnostic span when a dangerous
+/// trigger is used in various YAML list styles.
+#[test]
+fn test_dangerous_trigger_list_styles() -> anyhow::Result<()> {
+    insta::assert_snapshot!(
+    zizmor()
+        .input(input_under_test(
+            "dangerous-triggers/dangerous-block-list.yml"
+        ))
+        .run()?,
+    @"
+    error[dangerous-triggers]: use of fundamentally insecure workflow trigger
+     --> @@INPUT@@:3:5
+      |
+    3 |   - pull_request_target
+      |     ^^^^^^^^^^^^^^^^^^^ pull_request_target is almost always used insecurely
+      |
+      = note: audit confidence → Medium
+
+    4 findings (3 suppressed): 0 informational, 0 low, 0 medium, 1 high
+    "
+    );
+
+    insta::assert_snapshot!(
+    zizmor()
+        .input(input_under_test(
+            "dangerous-triggers/dangerous-flow-list.yml"
+        ))
+        .run()?,
+    @"
+    error[dangerous-triggers]: use of fundamentally insecure workflow trigger
+     --> @@INPUT@@:1:12
+      |
+    1 | on: [push, pull_request_target, workflow_call]
+      |            ^^^^^^^^^^^^^^^^^^^ pull_request_target is almost always used insecurely
+      |
+      = note: audit confidence → Medium
+
+    4 findings (3 suppressed): 0 informational, 0 low, 0 medium, 1 high
+    "
+    );
+
+    Ok(())
+}
+
+#[test]
 fn test_issue_comment() -> anyhow::Result<()> {
     insta::assert_snapshot!(
         zizmor()
@@ -27,11 +96,10 @@ fn test_issue_comment() -> anyhow::Result<()> {
             .run()?,
         @"
     error[dangerous-triggers]: use of fundamentally insecure workflow trigger
-     --> @@INPUT@@:1:1
+     --> @@INPUT@@:2:3
       |
-    1 | / on:
-    2 | |   issue_comment:
-      | |________________^ issue_comment is almost always used insecurely
+    2 |   issue_comment:
+      |   ^^^^^^^^^^^^^ issue_comment is almost always used insecurely
       |
       = note: audit confidence → Medium